refactor: update type imports and create logos.types.ts for better type organization
